c14h: return error when starting a transaction
I believe we are currently being bitten by #8 quite regularly. This is
exacerbated by not checking for an error by tx.Begin() (that's on me, I
presume), so we panic on a nil-pointer, instead of getting an error
message. Unfortunately, I still can't see a super clear-cut way to check
whether an error is caused by this issue and it shouldn't even *be* an
issue, as *sql.DB is a connection pool - so it should just handle this
transparently, IMO. But at least if we know the error message, we might
get a hint on how to handle that correctly.